Should be the final refactor
