+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/README.md
@@ -1,6 +1,10 @@
 # Using Motors
-**Note: Don't use this for drivetrains**
-Motors on the robot can be referenced with the `Motor` class. They can be controlled in autonomous OpModes with the SimpleMotorController and ComplexMotorController control planes. Keep in mind the SimpleMotorController and ComplexMotorControllers are both prototypes and will probably change later.
+
+**Note: Anything that looks like ~~this~~ is deprecated.**
+
+Motors on the robot can be referenced with the `Motor` class. They can be controlled in autonomous OpModes by themselves ~~or with the SimpleMotorController and ComplexMotorController control planes. Keep in mind the SimpleMotorController and ComplexMotorControllers are both prototypes and will probably change later.~~
+
+**Another Note: Everything below this is deprecated but i cant be asked to wrap every line with tildes**
 
 The simple control plane is best used for moving grouped motors at once as the simple control plane can only run one motor or motor group at a time, and must let the current group stop before running the next group.
